basic-restful-api/requirements.txt
Flask==1.1.1
google-cloud-datastore==1.10.0
requests==2.22.0